refactoring - decomposed code to severallogically named files
