+++ utils/utils.go
@@ -2,6 +2,7 @@ package utils
 
 import (
 	"net/http"
+	"path"
 	"strings"
 )
 
@@ -18,4 +19,13 @@ func ParseAuthHeader(w http.ResponseWriter, r *http.Request, authType, headerNam
 	login := creds[0]
 	password := creds[1]
 	return login, password, true
+}
+
+func ShiftPath(p string) (head, tail string) {
+	p = path.Clean("/" + p)
+	i := strings.Index(p[1:], "/") + 1
+	if i <= 0 {
+		return p[1:], "/"
+	}
+	return p[1:i], p[i:]
 }
